Feature/pn 4876 tbc qa fix test allarmi (#95)
* PN-4876 fix userAttributes error

* PN-4876

+++ src/test/java/it/pagopa/pn/cucumber/steps/SharedSteps.java
@@ -15,6 +15,7 @@ import it.pagopa.pn.client.b2b.pa.generated.openapi.clients.externalb2bpa.model.
 import it.pagopa.pn.client.b2b.pa.impl.IPnPaB2bClient;
 import it.pagopa.pn.client.b2b.pa.testclient.*;
 import it.pagopa.pn.client.web.generated.openapi.clients.externalUserAttributes.addressBook.model.LegalChannelType;
+import it.pagopa.pn.client.web.generated.openapi.clients.externalUserAttributes.addressBook.model.LegalDigitalAddress;
 import it.pagopa.pn.cucumber.utils.GroupPosition;
 import org.junit.jupiter.api.Assertions;
 import org.opentest4j.AssertionFailedError;
@@ -571,8 +572,11 @@ public class SharedSteps {
     public void vieneRimossaSePresenteLaPecDiPiattaformaDi(String user) {
         selectUser(user);
         try{
-            this.iPnWebUserAttributesClient.deleteRecipientLegalAddress("default", LegalChannelType.PEC);
-            logger.info("PEC FOUND AND DELETED");
+            List<LegalDigitalAddress> legalAddressByRecipient = this.iPnWebUserAttributesClient.getLegalAddressByRecipient();
+            if(legalAddressByRecipient != null && !legalAddressByRecipient.isEmpty()){
+                this.iPnWebUserAttributesClient.deleteRecipientLegalAddress("default", LegalChannelType.PEC);
+                logger.info("PEC FOUND AND DELETED");
+            }
         }catch (HttpStatusCodeException httpStatusCodeException){
             if(httpStatusCodeException.getStatusCode().is4xxClientError()){
                 logger.info("PEC NOT FOUND");
